# lib/api-zod/src/generated/types/fileNode.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { FileNodeType } from './fileNodeType';

export interface FileNode {
  name: string;
  path: string;
  type: FileNodeType;
  children?: FileNode[];
}