# lib/api-zod/src/generated/types/fileTree.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { FileNode } from './fileNode';

export interface FileTree {
  root: FileNode;
  projectName: string;
}