# lib/api-zod/src/generated/types/runInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { RunInputCommand } from './runInputCommand';

export interface RunInput {
  command: RunInputCommand;
}